requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pytest
pytest-asyncio
python-multipart
python-magic
httpx
aiofiles
